Refactor MOI wrapper into a single file
