+++ src/KNITRO.jl
@@ -6,6 +6,7 @@
 module KNITRO
 
 import Libdl
+import MathOptInterface as MOI
 import SparseArrays
 
 const _DEPS_FILE = joinpath(dirname(@__FILE__), "..", "deps", "deps.jl")
